A bit changes in calculation of payout ticket

+++ app/src/main/java/com/betkey/network/models/Rule.kt
@@ -19,5 +19,5 @@ data class Rule (
     var betslip: Double = 0.0,
 
     @SerializedName("bonus")
-    var bonus: Int? = null
+    var bonus: Double? = null
 )
